refactor: extract Coord into mod
